✨ Added Music to Both Scenes
Game Manager which handles the music for the scene
