+++ Darkspire/Assets/GD/My Assets/Prefabs/GameManager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e6e127fdf19a3f4458d1dd7e72e5a90d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
