+++ Darkspire/Assets/GD/My Assets/Scripts/Audio/SoundType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aee6d7a76a3a044409e2c16519a3d354
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
