+++ Darkspire/Assets/GD/My Assets/Scripts/Interactable/Lever/Lever.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 public class Lever : MonoBehaviour
@@ -37,6 +35,5 @@ public class Lever : MonoBehaviour
         {
             gate.StartMoving();
         }
-        SoundManager.PlaySound(SoundType.GATEOPEN);
     }
 }
