+++ Darkspire/Assets/GD/My Assets/Scripts/UI/MainMenu.cs
@@ -7,19 +7,10 @@ using UnityEngine.SceneManagement;
 public class MainMenu : MonoBehaviour
 {
 
-    public void Start()
-    {
-        MusicManager.Instance.PlayMusic("MainMenu");
-    }
-
     public void PlayGame()
     {
-        MusicManager.Instance.PlayMusic("Gameplay");
-        SceneManager.LoadScene("TestScene");
-        
+        SceneManager.LoadScene("MapPrototype");
     }
-
-    
 
     public void QuitGame()
     {
