+++ Darkspire/Assets/GD/Third Party Assets/Audio/Bacis/Music/GameplayMusic001.wav.meta
@@ -1,0 +1,23 @@
+fileFormatVersion: 2
+guid: 84793bceeb565e5499523886c38c2d31
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 7
+  defaultSettings:
+    serializedVersion: 2
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+    preloadAudioData: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
