+++ Darkspire/Assets/GD/Third Party Assets/Audio/Bacis/Music/music-for-trailers-mysterious-229751.mp3.meta
@@ -5,7 +5,7 @@ AudioImporter:
   serializedVersion: 7
   defaultSettings:
     serializedVersion: 2
-    loadType: 0
+    loadType: 2
     sampleRateSetting: 0
     sampleRateOverride: 44100
     compressionFormat: 1
